Move main function into lib.rs, to allow calling from other crates
